# app/Docs/apidocs.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "",
        "title": "Sistema de Gestion de Proyectos",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/proyecto": {
            "post": {
                "tags": [
                    "proyecto"
                ],
                "summary": "Agrega un nuevo proyecto",
                "description": "",
                "operationId": "postProyecto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Un objeto proyecto necesita agregar a proyecto",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Proyecto"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "pet"
                ],
                "summary": "Update an existing pet",
                "description": "",
                "operationId": "updatePet",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Pet object that needs to be added to the store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Proyecto"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
...
        "Proyecto": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string",
                    "example": "Majes Siguas II"
                },
                "fecha_inicio": {
                    "type": "string",
                    "example": "2019-12-10"
                },
                "fecha_fin": {
                    "type": "string",
                    "example": "2019-12-10"
                },
                "estado": {
                    "type": "string",
                    "example": "Aprobado"
                },
                "presupuesto": {
                    "type": "number",
                    "example": 900000
                }
            }
        },
...
    },
...
}